Changes to support DISTRO=lmp-base build without factory keys
The Foundries.io meta-lmp-base layer has a signing class which
errors if various key files and certificates are not present,
even when we are building an lmp-base image for development
which does not require signing.

The issue is in this class which tries to add hashes for the
files and errors if those files cannot be found

https://github.com/foundriesio/meta-lmp/blob/7657b41046ff745ce5fe61a64f615623ebb098e0/meta-lmp-base/classes/lmp-signing.bbclass#L56

I think the correct solution is probably only to call this
function if we're building a signed image but for now I have
worked around the issue by copying the bbclass and overriding
it by inheriting the modified copy in the failing recipes.

The modification generates a warning if the key files are not
found but no longer errors.

This now builds and tests out correctly with lmp-base

Signed-off-by: Alex J Lennon <ajlennon@dynamicdevices.co.uk>